Internal

Internal reference

Scenario tree

RPH.STreeNode — Type.

STreeNode

Node object of the ScenarioTree. Reference its father node id and its childs ids, along with the set of scenarios equivalent up to the depth (or stage) of the node.

source

Progressive hedging

RPH.ph_subproblem_solve — Function.

ph_subproblem_solve(pb, id_scen, u_scen, x_scen, μ, params)

TODO

source

Synchronous

RPH.randomizedsync_initialization! — Function.

randomizedsync_initialization!(z, pb, μ, subpbparams, printlev, it)

TODO

source

RPH.randomizedsync_subpbsolve — Function.

randomizedsync_subpbsolve(pb::Problem, id_scen::ScenarioId, xz_scen, μ, params)

Solve and return the solution of the subproblem 'prox(fs/μ) (xz_scen)' where 'fs' is the cost function associated with the scenario `idscen`.

source

Synchronous parallel

TODO

Asynchronous

RPH.AsyncSubproblemTask — Type.

AsyncSubproblemTask{T}

TODO

source

RPH.do_remote_work_async — Function.

do_remote_work_async(inwork::RemoteChannel, outres::RemoteChannel)

Solve and return the solution of the subproblem 'prox(fs/μ) (v_scen)' where 'fs' is the cost function associated with the scenario `idscen`.

source

RPH.randomizedasync_initialization_async! — Function.

randomizedasync_initialization_async!(z, pb, μ, subpbparams, printlev, it, work_channel, results_channel)

TODO

source

RPH.init_workers_async — Function.

init_workers_async(pb::Problem{T}, subpbparams::AbstractDict) where T<:AbstractScenario

TODO

source

Projections

RPH.get_averagedtraj — Function.

averaged_traj = get_averagedtraj(pb::Problem, z::Matrix{Float64}, id_scen::ScenarioId)

Compute and return the averaged trajectory defined by scenario id_scen over strategy z.

source

RPH.get_averagedtraj! — Function.

get_averagedtraj!(averaged_traj::Vector{Float64}, pb::Problem, z::Matrix, id_scen::ScenarioId)

Compute inplace in averaged_traj the averaged trajectory defined by scenario id_scen over strategy z. Note: this function has been fairly optimized. Apply changes with caution.

source

RPH.nonanticipatory_projection — Function.

x = nonanticipatory_projection(pb::Problem, y::Matrix{Float64})

Compute the projection x of y on the non-anticipatory subspace associated to problem pb.

source

RPH.nonanticipatory_projection! — Function.

nonanticipatory_projection!(x::Matrix{Float64}, pb::Problem, y::Matrix{Float64})

Store in x the projection of y on the non-anticipatory subspace associated to problem pb. Note: this function has been fairly optimized. Apply changes with caution.

source

Others

RPH.get_neighbydepth — Function.

get_neighbydepth(tree::STreeNode, scenid::ScenarioId)

Compute the leaves neighbooring scenid per level.

source

RPH.get_scenariodim — Function.

get_scenariodim(pb::Problem)

TODO

source

LinearAlgebra.dot — Function.

dot(pb::Problem, x::Matrix{Float64}, y::Matrix{Float64})

Compute the weighted scalar product between strategies x and y.

source