Internal

Internal reference

Scenario tree

RPH.STreeNode — Type.

STreeNode

Node object of the ScenarioTree. Reference its father node id and its childs ids, along with the set of scenarios equivalent up to the depth (or stage) of the node.

source

Progressive hedging

RPH.ph_subproblem_solve — Function.

ph_subproblem_solve(pb, id_scen, u_scen, x_scen, μ, params)

Solve subproblem associated with scenario id_scen, primal point x_scen, dual point u_scen and regularization μ.

source

Synchronous

RPH.randsync_subproblem_solve — Function.

randsync_subproblem_solve(pb::Problem, id_scen::ScenarioId, xz_scen, μ, params)

Solve and return the solution of the subproblem 'prox(fs/μ) (xz_scen)' where 'fs' is the cost function associated with the scenario `idscen`.

source

RPH.randsync_initialization! — Function.

randsync_initialization!(z, pb, μ, subpbparams, printlev, it)

Compute a first global feasible point by solving each scenario independently and projecting the global strategy obtained onto the non-anticipatory subspace.

source

Synchronous parallel

RPH.randpar_remote_func — Function.

randpar_remote_func(inwork::RemoteChannel, outres::RemoteChannel)

Solve and return the solution of the subproblem 'prox(fs/μ) (v_scen)' where 'fs' is the cost function associated with the scenario `idscen`.

source

RPH.randpar_initialization! — Function.

randpar_initialization!(z, pb, μ, subpbparams, printlev, it, work_channel, results_channel)

Compute a first global feasible point by solving each scenario independently and projecting the global strategy obtained onto the non-anticipatory subspace. Independent solves are distributed on available workers.

source

RPH.randpar_init_workers — Function.

randpar_init_workers(pb::Problem{T}, subpbparams::AbstractDict) where T<:AbstractScenario

Launch randpar_remote_func in every available worker with correct channels, and pass subpbparams to each worker.

source

RPH.randpar_terminate_workers — Function.

randpar_terminate_workers(pb, work_channel, remotecalls_futures)

Send a signal so that all workers return and wait that all workers do.

source

Asynchronous

RPH.randasync_remote_func — Function.

randasync_remote_func(inwork::RemoteChannel, outres::RemoteChannel, paramschan::RemoteChannel)

Solve and return the solution of the subproblem 'prox(fs/μ) (v_scen)' where 'fs' is the cost function associated with the scenario `idscen`.

source

RPH.randasync_initialization! — Function.

randasync_initialization!(z, pb, μ, subpbparams, printlev, it, work_channel, results_channel)

Compute a first global feasible point by solving each scenario independently and projecting the global strategy obtained onto the non-anticipatory subspace. Independent solves are distributed on available workers.

source

RPH.randasync_init_workers — Function.

randasync_init_workers(pb::Problem{T}, subpbparams::AbstractDict) where T<:AbstractScenario

Launch asyncpar_remote_func in every available worker with correct channels, and pass subpbparams to each worker.

source

RPH.randasync_terminate_workers — Function.

randasync_terminate_workers(pb, work_channel, remotecalls_futures)

Send a signal so that all workers return and wait that all workers do.

source

Projections

RPH.get_averagedtraj — Function.

averaged_traj = get_averagedtraj(pb::Problem, z::Matrix{Float64}, id_scen::ScenarioId)

Compute and return the averaged trajectory defined by scenario id_scen over strategy z.

source

RPH.get_averagedtraj! — Function.

get_averagedtraj!(averaged_traj::Vector{Float64}, pb::Problem, z::Matrix, id_scen::ScenarioId)

Compute inplace in averaged_traj the averaged trajectory defined by scenario id_scen over strategy z. Note: this function has been fairly optimized. Apply changes with caution.

source

RPH.nonanticipatory_projection — Function.

x = nonanticipatory_projection(pb::Problem, y::Matrix{Float64})

Compute the projection x of y on the non-anticipatory subspace associated to problem pb.

source

RPH.nonanticipatory_projection! — Function.

nonanticipatory_projection!(x::Matrix{Float64}, pb::Problem, y::Matrix{Float64})

Store in x the projection of y on the non-anticipatory subspace associated to problem pb. Note: this function has been fairly optimized. Apply changes with caution.

source

Others

RPH.get_neighbydepth — Function.

get_neighbydepth(tree::STreeNode, scenid::ScenarioId)

Compute the leaves neighbooring scenid per level.

source

RPH.get_scenariodim — Function.

get_scenariodim(pb::Problem)

Return the dimension of the vector representing a scenario.

source

LinearAlgebra.dot — Function.

dot(pb::Problem, x::Matrix{Float64}, y::Matrix{Float64})

Compute the weighted scalar product between strategies x and y.

source